requirements.txt

pillow
PyQt6
PyQt6-WebEngine
